Update README.md for rollup 4 hashes and handling regular static files
This addresses #114.

It also ensures that the whitenoise immutable file test also uses the default immutable file test.

`WHITENOISE_IMMUTABLE_FILE_TEST` is intended for usage when you're not using Django’s `ManifestStaticFilesStorage`. Just because we're using `django-vite` doesn't mean that we're not using `ManifestStaticFilesStorage`, so I **think**(?) we also need the default check.