Convert already_processed to a set()
We basically removed the recursive functionality
of already_processed in this PR:
https://github.com/MrBin99/django-vite/pull/109/files
We did that because the list concatenation was getting
so large for some users. But we can keep the benefits
of saving the recursive tags we've already_processed
(and do so without excessive bloat) by turning
already_processed into a set().